Fix build on travis
The OpenHIM test are failing with the latest secuirty update for
node.js. The February update included some changes to the way HTTP headers
are parsed and I think this is causing problems for us. This change
temporarily sets specific versions for the tests to run on that are the
latest releases before the secuirty update.

https://nodejs.org/en/blog/vulnerability/february-2016-security-releases/