Fix validation issues during rerun
Mongoose populates object in the schema with empty objects and empty
arrays by default. This causes the request object to be an empty object
even when not request property was even supplied. This then causes
Mongoose's validation to fail as it expects request objects to have a
.timestamp property.

Adding `required: false` to the schema seems to stop Mongoose from
populating it as an empty object.

https://github.com/jembi/openhim-console/issues/356