Add common root CAs to our https server
Right now we poison the CA list of our https server with only the
certificates that we trust from our keystore. This mean that TLS
verification will fail for any certificate that we don't have the full
certificate chain (up to the the root CA) for. It seems it is common for
root CAs not to give out their root certificates and rather rely on
intermediate certificates to prevent them from having to re-issue all
certificates if they happen to be compromised.

Thus, it is impossible to authenticate with the OpenHIM with a
certificate signed by one of the CAs where the root CA isn't available.
This means that for the most part only self signed chains can be used.

https://github.com/jembi/openhim-core-js/issues/655