Remove code to read metrics from statsd
It has been decided that we will no longer support reading metrics from
statsd as this is not often used and the current method of using the
mongo aggregation framework works really well. We no longer need to
support two different methods.

Metrics will still be reported to statsd as an option but no stats will
be read from statsd.