Include "from address" in email config
Nest the existing nodemailer config within a new email config section
and add the "from address" field. The code maintains backwards
compatibility with the old config format. This change is to allow for
the auth section of the nodemailer config to be left out in the case of
anonymous authentication.

See: OHM-398