Pin pem dependency version
A breaking change was introduced in a minor version bump which is
breaking the tests and causing the server to crash when updating the
cert. Pinning the version to the working one fixes the issue for now. At
some point the code should be updated to work with the latest version.