Reject new or updated clients or roles that conflict
ClientIDs and role name are used in multiple places interchangebly. We
should prevent the system from getting into a state where these conflict
so that we don't know if we are referring to a role or a client. This
commit addresses that issue by validating both clientID and role name.