Drop support for node versions below 7.6.0
This allows async/await to be used so that koa can take full advantage
of it. Node 8 will become the latest LTS version in October 2017 which
should be in time for the next major version of the OpenHIM to be
released. Support for the latest LTS should suffice and a workaround
using babel can be provided if needed.

OHM-467