Don't copy content-length header onto response
The content was being truncated or was too short in cases where the
upstream response included the content-length header and was using
different encoding to the OpenHIM response. Instead of using the value
of the content-length header from the upstream response, the OpenHIM
will now set its own content-length header when applicable.

OHM-516