Fix environment reading in config
Sometimes, depending on environment variable order, the NODE:ENV key in
nconf was being overwritten by a NODE environment variable. This was
happening particularly in workers. This is probably why nconf uses ':'
as a separator instead of '_' which we are using. Changing to not use
'_' would introduce breaking changes so this change rather reads the
environment directly from process.env.