Clean up transaction status handling
Centralise all of the logic for setting the transaction's status within
the setFinalStatus function so that there is a single place to record
the outcome of the transaction. Also clean up the async flow and error
handling for setting the final status - previously some errors were
silenced and others could cause unhandled promise rejections.

OHM-684