Replace the "mongoose-pacth-history" dependency with a version Jembi forked
The dependency has a bug which doesnt cater for the $push/$pull mongo operators
A PR has been opened to the maintainers: https://github.com/codepunkt/mongoose-patch-history/pull/30

With the updated dependency, a bug has been noticed with the reference to "updatedBy" that was set by "virtuals".
This has been replaced by adding the "updatedBy" property to the Channels schema as it has value in storing it.

OHM-705