Fixed incorrect updating of transaction route
When updating a transaction with one non-primary route, it ended up triggering the mongo update twice and adding a duplicated route to the routes array.
Im not exactly sure how this works, but investigating the issue lead me to the below thread that mentions the issue stems from a mixed use of async/await and callbacks.
https://stackoverflow.com/questions/49372428/mongoose-push-keeps-adding-two-entries

Removing the intial return of the mongo update and letting it be handled with the function callback solved the duplication issue.

OHM-705