Remove the clearing of the database from the script
This was causing race conditions with some of the files which were already connected to a mongo connection and then not having a reference to an existing collection. This was picked up in the LogsAPITests were the tests were failing intermittently.
Replaced the inclusion of the entire utils script with just the required config script to override the router httpPort

OHM-732