extract the method for retrieving
The functionality for storing and retrieving of the transaction bodies is held in the contentchunk.js file. This commit moves the methods for retrieving the bodies from gridfs into the contentChunk.js file. The retrieveTransactionBody method has also been modified. It no longer takes in the database handle as an argument. The tests have also been moved and modified

OHM-763