change the Route Orchestrations in schema
The orchestration definition has been changed in the transaction schema. The route definition has the property orchestrations of orchestraion type. This commit will ensure the tests pass by setting the route orchestraions to use the old orchestration type. The routes definition will be modified to use the new Orchestration type when routes are modified (in the next sprint)

OHM-772