Remove bodies from orchestrations during body culling
Transactions bodies contained within the orchestrations can be large
chunks of data and for whatever reason a user wants that data to no
longer be stored (Due to POPI, space issues, whatever) then the body
culling feature will remove the bodies of transactions older than a
chosen date.

We need to ensure that the body is deleted out of the GridFS data base
as well as removing the BodyId from the transaction.
The body chunk needs to be removed from the DB to prevent a build up of
orphaned data within Mongo.
The bodyID needs to be removed from the transaction to prevent tryign to
lookup data that no longer exists.

OHM-774 OHM-691