modify the function for sending requests
some of the statements had to be adjusted after a merge had been done. The timestamp names have been changed to match the transaction schema timestamp names. The winston logger function is now used instead of the console.log. This for consistency as winston is whats used thoughout. The logic for checking the request method before writing chunks to the request object has also been changed so that its only executed once, not on every chunk as was.

OHM-790