refactor the rawBodyReader function
Refactoring of this function has been done as some of the logic checks for executing some statements was not necessary. There was logic to check if variables defined in the function existed, and to set them to zero once the function finishes. This is reduntant as the variables are local to the function and are redefined every time the function is called. The bucket variable has been moved outside of the function so that the gridfs bucket is only created onc not on every request as was. Logic has also been added to check the request method. THis is so we only add the bodyId property when the request is not a get or delete