change the update objects used for updating the transaction
The object used for updating the transaction when there is an error had a 'response.message' property which does not exist in thransaction schema. The property has been removed and the error property of the transaction will be updated instead. The update object used for updating the transaction's response has also had the 'repsonse.message' removed. it was being updated with the status message, which is not of much use as we already have the response status code