Remove request bodies when rerunning transactions
Prior to this change, request body was retruned from GridFS and
added to the transaction before submitting it for a rerun.

This change removes the body from the transaction before it is
submitted to rerun. For transactions that originally had a body,
the body should already be stored in GridFS and only the
bodyId (GridFS fileId) is needed to retrieve it. The bodyId is
passed as a custom header and then submitted on the rerun port.

OHM-807