create a function for sending an http request
This is the function that will be used for sending secondary routes requests. It uses some of the functionality that is used in the function for sending the primary route requests. Refactoring will be done later on to apply the DRY principle. The function has been added to the 'sendRequest' function, which is the wrapper for the functions that send the requests. A flag has been added to use the function for secondary routes when route is secondary.

OHM-800