add the initiation of a response on error
The metrics were not being recorded when there was an error on the primary route request. The erroring was resulting in the response timestamp not being set, as it is set on the response stream 'data' event. THe response timestamp is needed for the metrics. Functionality that sets the response timestamp will now be executed even when there is an error