Allow makeStreamingRequest to receive a null input stream (request body)
Ease of use - if no input stream is supplied (this would normally
contain the request body to send), the function would behave as
if an empty body is to be sent (treat transaction as a GET request).

This prevents the function from erroring out if no stream is provided.
An input stream is always required to hang event handlers off of.

OHM-809