change the calling of the function to update the status
The function that updates the status of the transaction was being called every time an error would occur, but the functionality was updating the koa context only and not throwing an error, resulting in the "setFianalStatus" function being called many times. The function is now only called after all the prommises that send the routes' requests hvae been resolved.