Bug fix: Responses in Mediator format was failing the transaction
This was due to the change to read an incoming stream instead of a full payload.
This change takes the reponse.body stream and builds up a string with all the supplied chuncks before setting the koa response.
We need to read in the entire payload here to process the mediator medatadata being supplied in the response