add funtionality that updates the transaction error property
When an error occurs during the routing of the request, the koa context is updatedwith the error and the transaction is updated with the response (status and timestamps). The error property of the transaction will now also be updated. This also fixes the some of the tests that where failing

OHM-820