fix the autoRetry functionality
The transaction autoretry functionality was not working as should. The number of retry attempts was not being recorded, and this was resulting in the "max number of retries set" being ignored. This was resulting in continous retry process. The koa mddleware that sets the number of retries property has been moved such that it happens before the transaction is recorded in the database.

OHM-824