fix the setting of the status from a mediator response body
The functionality that sets the status of the transaction from the response body of a mediator was not working. The function that sets the koa response (setKoaResponse) takes the ctx object and the response object. The body of the response was a stream and this has been changed. It is now a string buffer and the status is now set based on the status property of the response body. The 'setKoaResponse' function was also being called twice, and it now called only once. This fixes some of the tests that were failing as well

OHM-825 OHM-820