set the koa context 'mediatorResponse' property
In fixing the setting of the transaction statuscode from a mediator's response body, the statement for setting of the ctx mediatorResponse was left out. The mediatorResponse is set to the json parsed response (from the mediator), and is then used in the setting of the final status. This is for setting the transaction status (string) not the status code

OHM-825