Add feature: streamingRouter can aggregate response body
Prior to this change, the streamingRouter received the response from
downstream as a series of chunks. This meant that string, regex, etc
functions could not be applied to the response as all data hadn't
been collected into a single buffer. Hence, url rewriting could not
be applied.

This change adds an option to the streamingRouter that differentiates
between receiving a streaming response, or collecting all the chunks
into a buffer(string).

In 'collectResponseBody' mode, all chunks are
collected and once all are received, the buffer is passed to an
event handler to permit url rewriting (or any other post-processing)
to take take place. On return, the full buffer is stored in GridFS
and sent back upstream to the calling client [as a stream,not string]

In 'streaming mode' mode, all chunks are persisted to GridFS and sent
upstream to the the client, as they are received

OHM-829