change the functionality for storing mediator responses
when a mediator sends back the response, the whole response object is streamed directly into gridfs. We should only store the response body after parsing the mediator response. This way only the transaction response body will be shown on the front end and not the mediator response object. This is how the openhim was before the refactoring. This commit fixes the issue.

OHM-813