change the updating of the transaction
The status and the response body Id were being updated in the function initiateResponse from the koa context, and this is before the body id and the status had been set. The two properties are set in the 'setKoaResponse' function which gets called after the initiateResponse function. THe updating has been moved to the function 'completeResponse' which gets called after the 'setKoaResponse'

OHM-813