fix routes integration tests
One test was failing beacuse the test channel responseBody property had not been set . When its not set, the bodyId is set null. There is an ssertion for the bodyId which was being set to null. THe test channel responseBody property has been set and the test is now passing.

OHM-820