Make routing conditional based on authenticated user
Prior to this change, routing middleware would execute if the
authorisation module mathes a user to a channel. Due to the order
that authorisation and requestMatching occur in the streaming
version, this is no longer sufficient.

This change breaks the middleware chain in the routing module if,
by this point, a user hasn't been authorised or we want to revoke a
user's authorisation. This means we have up to the execution of the
routing middleware to invalidate a user's authorisation.

OHM-831