Fixed broken metrics tests
With some of the new changes we have made to the routing engine, we have broken the asserts within the metrics test file. We have introduced a new request/response property for timestampEnd which contains the timestamp of the request/response completed. The tests were not updated to contain the new property which resulted in the metris never being created for the test case

OHM-820