Handle content matching in the streamingReceiver (old rawBodyParser)
Prior to this change, matching on the body content of the request
happened before the body was processed. With the addition of stream
handling, the entire request body needs to be loaded in memory for
string/regex scanning to be effective.

This change creates two "modes" of reading in and processing the
request:
(1) fully-streamed
  - where each chunk of the request body is routed downstream as it
    is received
  - if there is a content-match (string/regex/json/xml), the full
    body is stored in GridFS and routed downstream (as a stream)
(2) collected -
  - to enable string/regex scanning of the request body, all the
    body chunks are aggragated into a single buffer for scanning
  - If there is no content match after scanning, the transaction
    is de-authorised (and the middleware is completed; no storage
    to GridFS and no transaction is persisted)

OHM-831