await the collection of openhim mediator response
The response from an openhim mediator comes in a special format, and it has to parsed so that the response body, status, etc can be extracted. This means that it has to be stored in memory. The method that collects the chunks of the mediator's response is asynchronous and it has to be waited on. This is so we can set the right status and send back the body only (without the other fields on the mediator's response) to the client

OHM-824