remove initiation of response on failure
whenever there is an error, the method updateWithError is invoked. Calling initiateResponse is not necessary as updateWithError does the updates that initiateResponse does. Calling the two methods at once was causing a race condition. At times the error was not being updated, resulting in some tests which have assertions on the error to be inconsistent

OHM-824