fix up tests
The property 'responseBody' of the channel has been set to true. This is because, at the moment when the property is false or undefined, the app does not send a response to the client. This sending of the response issue (when channel's responseBody property is set to false) has been addressed on another branch. This is a temporary fix to get the tests to pass.

OHM-824