modify the logic for setting the autoRetry
The logic had to be modified as all transactions that failed were being queued up for auto Retry, even when the channel is set not to auto retry a failed transction. The queueing of the transactions to be retried and the auto retries should only occur when the channel's property "autoRetryEnabled" is true. This way only transactions that fail after the autoRetry is enabled are rerun

OHM-824