add tcp channel properties to the context
The properties 'isTcpChannel' and 'tcpChannelHasHttpRoute' are now added to the context for tcp channel requests. These properties are used in the streaminReceiver middlware, which comes after this middleware, to determine whether the middleware is bypassed or not. The streamingReceiver is used by the http routes. This will cater for cases when we route from tcp to http. These properties will be set to true for that case and then the streaming middleware is used

OHM-907