modify the logic for setting collectBody
This has been doen because with the logic that was there the collectBody property was always true even when it had to be false. The collectBody is used to detemine whether the request body should be buffered in memory before anything else happens. This is for cases when we are doing request matching on content.

OHM-907