fix logic for handling cases where transaction does not exist
The logic for handling cases where the transaction does not exist (in the setFinalStatus method) has been modified. The callback that is returned by this function was being called before checking whether the transaction exists. The callback was being called with a value of null for the transaction which causes errors as the callback function tries to extract an id from null

OHM-907