Apply the correct date types
For the mongoose modal fitler parameters java date objects are required
for the search function. This was accidentally converted to a momentjs
date object which doesn't work.

The momentjs date object is however easier to manipulate and is
therefore useful for defining the date and for use in the templating of
the reports.

The use of the two different date objects needs to be used with care

OHIE-278