Fixed test startup scripts
The babel register package has been updated and required a slightly
different implementation.

The unit and integration scripts were also updated to work with latest
babel/register.

The mocha timeout had to be increase to 10 seconds to cater for some
slow test cases. This is not a great fix and requires investigation. But
it allows for the dependency upgrade which addresses security issues.
I've logged a ticket to investigate the timeout issue.

https://jembiprojects.jira.com/browse/OHM-923

OHIE-277