Clone and split Moment date object
The format of the moment date is different in the two places it is used.
For display in the email a human readable format is needed. For the
filter and within the email links the iso string format is needed.

To get the two formats the date object needs to be cloned as the object
is mutated when the format is changed and the object asigns used in the
filter and email reference the original object therefore any changes
would have affected all object assigns.

OHIE-277