Fix report email template links
The channel link took the user to the channel settings view. This is not
a useful view. This should rather take the user to the channel metrics
view as that is the common flow from the report. One wouldn't be looking
for channel settings after looking at the report.

The `Total transactions` link also shouldn't take the user to the
channel metrics. The total value indicates the total number of
transaction for the period of the report. The metrics page does not show
that. This is the job of the transactions page and the appropriate
filters (namely: channelID, period start and period end)

OHIE-277