Fixed broken TCP-> TLS tests
Fixed afew broken tests

With changes made in Node 12, a new TLS version is used which caused some tests to break due to incorrect cipher usage.
Added a .env.test file to load environment variables when testing. This was due to an addition NODE_OPTIONS env variable that was adding which is starting to bloat the start line

OHM-938