Update README documentation
Updated references to node 12 as this is our latest supported LTS version. A newer version of MongoDB can now be used.
Removed the "deployments" section as this is no longer valid. The openhim-docker repository handles the building of the images and isnt trigger on pushing new code to the master branch

OHM-941