Add new function for starting the API http server
The current logic is missing the some key logic for starting the API
HTTP server. This commit adds a new function for starting the server and
contains logic to ensure that the default admin user is created.

This commit is in response to a review request by @BMartinos at https://github.com/jembi/openhim-core-js/pull/1073#discussion_r368836220

Related Tickets:
 - OHIE-287