Set enable JWT Auth to false
Enabling a new auth mechanism by default might cause setup issues with
unsuspectign users as previously the OpenHIM only supported Basic Auth.
The default config of JWT would not work and would throw errors in the
logs with every request.
Setting some proper default config options for JWT also is not an option
as we currently don't have an easy way to change this config as we have
with Basic Auth.

TRACE-126