Simplify client custom token id update logic
The Client API will not update fields if they are undefined. It will
however update a field if it is null.

This update method can be taken advantage of to reduce the logic needed
to know when to update the custom token id.

If the custom token ID has a value this value will not be sent to the
requester for security. However the requester will receive a field
indicating that the custom token field field was set.

If the custom token id is updated then the api will receive the new
value to update it to either a string or null.

Else the API will receive undefined for the field and it won't update
the existing token ID

TRACE-140