Remove killing of secondary routes when primary fails
If we kill the secondaries we have no idea if they completed or not. This is inconsitent, a better apporach is to let them run to completion. In the future we could add in an option to a channel to allow secondary route to only start running if the primary completes.

OHM-1047