Allow re-runs to function when body storage is turned off
If we have a body from a previous transaction but bodies are turned off now in the channel. Then we should still be able to re-run that transaction. This change removes and incorrect check that makes the re-run request go down the path of the GET request when it infact has a body.

OHM-1046