Compact transactions collection during GridFS migration
MongoDB does not give back space used by deleted documents. During this
GridFS migration the transactions collection will lose most of it's data
which will be stored in the `fs.chunks` collection as we make use of
GridFS.
The transactions collection is where most data from the openhim is
generated - therefore doubling the data here would likely cause near
doubling of the DB storage size which is not ideal nor always possible.
Therefore we run the compact collection command after every migration
batch in order to keep the required DB space to a minimum.

OHM-1054