fix deleting of autoretries
The logic for deleting auto retries has been fixed. When a retry is done, it is deleted but when multiple are processed (in the agenda job), the logic that existed was only deleting a single retry out of the processed ones, and this would result in more retries happening than should. The agenda job runs after a minute.

OHM-1049