Fix importing of polling channels
We don't register polling channels when import metadata from a file. This is because we go directly to the database and don't use our programmatic API that was designed to cover these sorts to nuances. This is why builting interfaces to DB access (or any external system) is important. It allows you to put all the handling code in one place.

Eventually we should re-work the metadata import to rather use the API for all objects that are imported. This, however, is a much larger task.