Add endpoint for creating rerun tasks for bulk reruns
This is a more efficient way of creating the bulk rerun tasks. Before the openhim console had to fetch transcation ids and then send them back to the openhim core for the creation of rerrun tasks. The openhim console only has to specify the query parameters for the transactions to be rerun and the core will handle the creation of the tasks.