Replace user id with user email in passport
The email address of a user in now being used as the link between a user and their passport instead of the user's `_id`. The email is unique for each user, and is something that set where as the user id is created at the database level, and one would have to retrieve the user's id to link it a passport, which would make the metadata importation a bit complicated. Creating of a user and their passport can now be done at the same time as one no longer needs the user's email for linking it to a passport.

PLAT-719